解决gcc编译警告

+++ Source/Server/CloudSleepServer/logger.cpp
@@ -54,7 +54,7 @@ char* GetTimeStampStr()
     static char buffer[256] = { };
 
     snprintf(buffer, 256, "%04d-%02d-%02d %02d:%02d:%02d:%03d", tm->tm_year + 1900, 
-        tm->tm_mon + 1, tm->tm_mday, tm->tm_hour, tm->tm_min, tm->tm_sec, msec);
+        tm->tm_mon + 1, tm->tm_mday, tm->tm_hour, tm->tm_min, tm->tm_sec, (int)msec);
     return buffer;
 }
